+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - Removed named parameters from `Puppetfile#initialize` to support older Rubies.
+- Moved all IO to `Puppetfile::CLI`.
 
 ### Fixed
 - Proper status for `kept at ...` messages.
 - Return earlier on `:matched` status so that no re-ordering of parameters can occur.
+- Various CLI logic issues.
 
 ## [0.2.0] - 2017-09-25
 ### Added